requirements.txt
spacy==2.0.18
tensorflow-gpu==1.15.0
numpy==1.15.4
sklearn
